Remove default React files and delete workspace folder
