+++ frontend/public/chatbot_page/chatbot.css
@@ -1,0 +1,160 @@
+:root{
+  --color-primary:#002D56;
+  --color-accent:#8D7150;
+  --color-secondary:#146E7A;
+}
+
+/* ─── Global & Background ─── */
+body{
+  margin:0;
+  font-family:'Segoe UI',sans-serif;
+  background-image:linear-gradient(120deg,#002D56 0%,#146E7A 50%,#8D7150 100%);
+  background-size:400% 400%;
+  background-position:0% 50%;
+  animation:idleWave 20s ease-in-out infinite;
+  color:#fff;
+  display:flex;
+  justify-content:center;
+  align-items:center;
+  height:100vh;
+}
+
+/* Idle 파동 */
+@keyframes idleWave{
+  0%  {background-position:0% 50%;}
+  50% {background-position:100% 50%;}
+  100%{background-position:0% 50%;}
+}
+
+/* Surf 파도 */
+body.surf{
+  animation:surfWave 2s ease-in-out infinite;
+}
+
+@keyframes surfWave{
+  0%  {background-position:50% 100%;}
+  50% {background-position:50%   0%;}
+  100%{background-position:50% 100%;}
+}
+
+/* ─── Chat Wrapper ─── */
+.chat-wrapper{
+  width:960px;
+  height:700px;
+  background:#fff;
+  border-radius:16px;
+  overflow:hidden;
+  display:flex;
+  flex-direction:column;
+  box-shadow:0 8px 24px rgba(0,0,0,0.3);
+}
+
+.chat-header{
+  background:var(--color-secondary);
+  padding:20px;
+  color:#fff;
+  font-size:1.5rem;
+  font-weight:bold;
+  display:flex;
+  justify-content:space-between;
+  align-items:center;
+  position:relative;
+}
+
+/* ─── FAQ ─── */
+.faq-menu{
+  position:relative;
+  cursor:pointer;
+  font-size:1rem;
+  background:rgba(255,255,255,0.15);
+  padding:8px 12px;
+  border-radius:8px;
+  transition:background .2s ease;
+}
+.faq-dropdown{
+  display:none;
+  position:absolute;
+  top:60px;
+  right:20px;
+  background:#fff;
+  color:#333;
+  border-radius:8px;
+  box-shadow:0 4px 16px rgba(0,0,0,0.2);
+  z-index:100;
+  min-width:200px;
+  overflow:hidden;
+}
+.faq-dropdown.show{display:block;}
+.faq-dropdown div{
+  padding:12px 16px;
+  border-bottom:1px solid #eee;
+  cursor:pointer;
+  transition:background .2s ease;
+}
+.faq-dropdown div:hover{background:#f5f5f5;}
+/* 나가기 버튼 특화 스타일 */
+.exit-btn{font-weight:bold;}
+.exit-btn:hover{background:#ffecec;color:#e00;}
+
+/* ─── Chat area ─── */
+.chat-body{
+  flex:1;
+  padding:24px;
+  overflow-y:auto;
+  background:#f1f1f1;
+}
+.message{margin-bottom:16px;display:flex;}
+.message.user .bubble{margin-left:auto;background:var(--color-primary);color:#fff;}
+.message.bot  .bubble{margin-right:auto;background:var(--color-accent);color:#fff;}
+
+.bubble{
+  padding:14px 20px;
+  border-radius:20px;
+  max-width:60%;
+  font-size:1rem;
+  line-height:1.5;
+  word-wrap:break-word;
+}
+
+/* ─── Loader (3 blinking dots) ─── */
+.loader{display:flex;gap:6px;align-items:center;}
+.loader span{
+  width:8px;height:8px;border-radius:50%;
+  background:#fff;opacity:.25;
+  animation:blink 1.2s infinite ease-in-out;
+}
+.loader span:nth-child(2){animation-delay:.2s;}
+.loader span:nth-child(3){animation-delay:.4s;}
+
+@keyframes blink{
+  0%,80%,100%{opacity:.25;}
+  40%{opacity:1;}
+}
+
+/* ─── Input ─── */
+.chat-input{
+  display:flex;
+  border-top:1px solid #ccc;
+  background:#f9f9f9;
+  padding:16px;
+}
+.chat-input input{
+  flex:1;
+  border:1px solid #ccc;
+  padding:12px 16px;
+  font-size:1rem;
+  border-radius:8px;
+  outline:none;
+}
+.chat-input button{
+  margin-left:12px;
+  background:var(--color-secondary);
+  border:none;
+  color:#fff;
+  padding:12px 24px;
+  font-size:1rem;
+  border-radius:8px;
+  cursor:pointer;
+  transition:background .2s ease;
+}
+.chat-input button:hover{background:#0f5c67;}
