+++ frontend/public/chatbot_page/chatbot.html
@@ -1,0 +1,40 @@
+<!DOCTYPE html>
+<html lang="ko">
+<head>
+<meta charset="UTF-8">
+<meta name="viewport" content="width=device-width, initial-scale=1.0">
+<link rel="stylesheet" href="chatbot.css">
+<title>HUBOT</title>
+</head>
+<body>
+<div class="chat-wrapper">
+  <!-- Header -->
+  <div class="chat-header">
+    <span>한국외국어대학교 HUBOT</span>
+    <div class="faq-menu" id="faq-toggle">
+      자주 묻는 질문 ▾
+      <div class="faq-dropdown" id="faq-dropdown">
+        <div>2025학번 졸업학점은 몇 학점인가요?</div>
+        <div>전공필수 과목은 무엇인가요?</div>
+        <div>정보통신공학과 실습실은 어디에 있나요?</div>
+        <div class="exit-btn" id="exit-btn">나가기</div>
+      </div>
+    </div>
+  </div>
+
+  <!-- Body -->
+  <div class="chat-body" id="chat-body">
+    <div class="message bot">
+      <div class="bubble">안녕! 무엇을 도와줄까?</div>
+    </div>
+  </div>
+
+  <!-- Input -->
+  <form class="chat-input" id="chat-form">
+    <input type="text" id="chat-input" placeholder="메시지를 입력하세요…" autocomplete="off">
+    <button type="submit">보내기</button>
+  </form>
+</div>
+<script src="chatbot.js"></script>
+</body>
+</html>
