+++ frontend/public/index.html
@@ -1,43 +1,63 @@
 <!DOCTYPE html>
-<html lang="en">
+<html lang="ko">
   <head>
-    <meta charset="utf-8" />
-    <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <meta name="theme-color" content="#000000" />
-    <meta
-      name="description"
-      content="Web site created using create-react-app"
-    />
-    <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
-    <!--
-      manifest.json provides metadata used when your web app is installed on a
-      user's mobile device or desktop. See https://developers.google.com/web/fundamentals/web-app-manifest/
-    -->
-    <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
-    <!--
-      Notice the use of %PUBLIC_URL% in the tags above.
-      It will be replaced with the URL of the `public` folder during the build.
-      Only files inside the `public` folder can be referenced from the HTML.
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>링크 이동 버튼 예제</title>
+    <script>
+      function goTo(url) {
+        window.location.href = url;
+      }
+    </script>
+    <style>
+      body {
+        font-family: Arial, sans-serif;
+        padding: 20px;
+      }
 
-      Unlike "/favicon.ico" or "favicon.ico", "%PUBLIC_URL%/favicon.ico" will
-      work correctly both with client-side routing and a non-root public URL.
-      Learn how to configure a non-root public URL by running `npm run build`.
-    -->
-    <title>React App</title>
+      button {
+        display: block;
+        margin: 10px 0;
+        padding: 12px 20px;
+        font-size: 16px;
+        border: none;
+        background-color: #3498db;
+        color: white;
+        border-radius: 6px;
+        cursor: pointer;
+        transition: background-color 0.3s;
+      }
+
+      button:hover {
+        background-color: #2980b9;
+      }
+    </style>
   </head>
   <body>
-    <noscript>You need to enable JavaScript to run this app.</noscript>
-    <div id="root"></div>
-    <!--
-      This HTML file is a template.
-      If you open it directly in the browser, you will see an empty page.
+    <h1>페이지 이동 버튼</h1>
 
-      You can add webfonts, meta tags, or analytics to this file.
-      The build step will place the bundled scripts into the <body> tag.
+    <button onclick="goTo('http://localhost:3000/main_page/main_page.html')">
+      http://localhost:3000/main_page/main_page.html
+    </button>
 
-      To begin the development, run `npm start` or `yarn start`.
-      To create a production bundle, use `npm run build` or `yarn build`.
-    -->
+    <button onclick="goTo('http://localhost:3000/login_page/login_page.html')">
+      http://localhost:3000/login_page/login_page.html
+    </button>
+
+    <button
+      onclick="goTo('http://localhost:3000/notice_page/notice_page.html')"
+    >
+      http://localhost:3000/notice_page/notice_page.html
+    </button>
+
+    <button
+      onclick="goTo('http://localhost:3000/completion_system_page/com_sys_page.html')"
+    >
+      http://localhost:3000/completion_system_page/com_sys_page.html
+    </button>
+
+    <button onclick="goTo('http://localhost:3000/sub_page/sub_page.html')">
+      http://localhost:3000/sub_page/sub_page.html
+    </button>
   </body>
 </html>
